Fix #327 - Move linux dependencies into a subdir
