+++ src/linux/testdata/cpuinfo-aarch64.txt
