+++ src/linux/testdata/cpuinfo-x86_64.txt
