+++ src/linux/testdata/stat.txt
